aws-mysql56-ubuntu-python/requirements.txt
mysql-connector-python==8.0.22
pandas==1.2.0
SQLAlchemy==1.3.22
pkg-resources==0.0.0
